# tp4-ecommerce/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shopecart E-commerce API Documentation",
        "description": "Documentation des points d'accès (endpoints) de l'API Shopecart.",
        "contact": {
            "email": "contact@shopecart.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de l'API locale"
        }
    ],
    "paths": {
...
        "/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get user's cart",
                "operationId": "86f867f72e674aacd82942dcf38ba77e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "CReate Cart",
                "operationId": "c0e8ca78699ccc07858444da702d3ac5",
                "responses": {
                    "200": {
                        "description": "ICReate cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/items/{cartItem}": {
            "put": {
                "tags": [
                    "Cart"
                ],
                "summary": "Update cart item quantity",
                "operationId": "4821dca79c7ed9c583c1598cb450fc5d",
                "parameters": [
                    {
                        "name": "cartItem",
                        "in": "path",
                        "description": "Cart Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cart item updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove item from cart",
                "operationId": "8d114a15678817e115627d0eace73231",
                "parameters": [
                    {
                        "name": "cartItem",
                        "in": "path",
                        "description": "Cart Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed from cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cart/clear": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear the cart",
                "operationId": "c2cf94402630acad1cc6d317ec806693",
                "responses": {
                    "200": {
                        "description": "Cart cleared successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cart": {
                "title": "Cart",
                "description": "Modèle de données pour un panier d'achat.",
                "properties": {
                    "id": {
                        "description": "ID unique du panier.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur associé à ce panier.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "items": {
                        "description": "Liste des articles dans le panier.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItem"
                        }
                    },
                    "total_price": {
                        "description": "Prix total du panier.",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "user_id": 101,
                    "items": [],
                    "total_price": 49.99
                }
            },
            "CartItem": {
                "title": "CartItem",
                "description": "Modèle de données pour un article dans un panier.",
                "properties": {
                    "product_id": {
                        "description": "ID du produit.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "quantity": {
                        "description": "Quantité de ce produit dans le panier.",
                        "type": "integer"
                    },
                    "unit_price": {
                        "description": "Prix unitaire au moment de l'ajout.",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object",
                "example": {
                    "product_id": 5,
                    "quantity": 2,
                    "unit_price": 24.99
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez le jeton Bearer obtenu après la connexion.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}